+++ internal/actions/mocks/client.go
@@ -24,6 +24,7 @@ type TestData struct {
 	NameOfContainerToKeep   string
 	Containers              []t.Container
 	Staleness               map[string]bool
+	RenameContainerError    error
 }
 
 // TriedToRemoveImage is a test helper function to check whether RemoveImageByID has been called
@@ -62,7 +63,7 @@ func (client MockClient) StartContainer(_ t.Container) (t.ContainerID, error) {
 
 // RenameContainer is a mock method
 func (client MockClient) RenameContainer(_ t.Container, _ string) error {
-	return nil
+	return client.TestData.RenameContainerError
 }
 
 // RemoveImageByID increments the TriedToRemoveImageCount on being called