+++ internal/actions/mocks/container.go
@@ -19,7 +19,7 @@ func CreateMockContainer(id string, name string, image string, created time.Time
 		ID:      id,
 		Image:   image,
 		Name:    name,
-		Created: created.String(),
+		Created: created.Format(time.RFC3339Nano),
 		HostConfig: &dockerContainer.HostConfig{
 			PortBindings: network.PortMap{},
 		},
@@ -56,7 +56,7 @@ func CreateMockContainerWithImageInfoP(id string, name string, image string, cre
 		ID:      id,
 		Image:   image,
 		Name:    name,
-		Created: created.String(),
+		Created: created.Format(time.RFC3339Nano),
 		Config: &dockerContainer.Config{
 			Image:  image,
 			Labels: make(map[string]string),
@@ -85,7 +85,7 @@ func CreateMockContainerWithConfig(id string, name string, image string, running
 			Running:    running,
 			Restarting: restarting,
 		},
-		Created: created.String(),
+		Created: created.Format(time.RFC3339Nano),
 		HostConfig: &dockerContainer.HostConfig{
 			PortBindings: network.PortMap{},
 		},
@@ -119,7 +119,7 @@ func CreateMockContainerWithLinks(id string, name string, image string, created 
 		ID:      id,
 		Image:   image,
 		Name:    name,
-		Created: created.String(),
+		Created: created.Format(time.RFC3339Nano),
 		HostConfig: &dockerContainer.HostConfig{
 			Links: links,
 		},