+++ pkg/sorter/sort.go
@@ -22,9 +22,9 @@ func (c ByCreated) Less(i, j int) bool {
 		t1 = time.Now()
 	}
 
-	t2, _ := time.Parse(time.RFC3339Nano, c[j].ContainerInfo().Created)
+	t2, err := time.Parse(time.RFC3339Nano, c[j].ContainerInfo().Created)
 	if err != nil {
-		t1 = time.Now()
+		t2 = time.Now()
 	}
 
 	return t1.Before(t2)